lendex-token: move DisplayAmount to its own file
